// src/app/api/models/game-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { GameCommentDto } from '../models/game-comment-dto';
import { GameLeaderboardDto } from '../models/game-leaderboard-dto';
import { GameRatingDto } from '../models/game-rating-dto';
import { GameTagDto } from '../models/game-tag-dto';
export interface GameDto {
  averageRating?: number;
  category?: 'ACTION' | 'ADVENTURE' | 'ARCADE' | 'BOARD' | 'CARD' | 'CASINO' | 'RPG' | 'STRATEGY';
  comments?: Array<GameCommentDto>;
  creationDate?: string;
  description?: string;
  difficultyLevel?: 'EASY' | 'MEDIUM' | 'HARD';
  id?: number;
  imageUrl?: string;
  isActive?: boolean;
  lastModifiedDate?: string;
  leaderboardEntries?: Array<GameLeaderboardDto>;
  minAge?: number;
  name?: string;
  playCount?: number;
  ratings?: Array<GameRatingDto>;
  tags?: Array<GameTagDto>;
  url?: string;
}